requirements.txt
mediapipe
opencv-python
Pillow